+++ ios_base_foundation/Classes/Extension/UIKit/UIImageExtensions.swift
@@ -139,11 +139,6 @@ public extension UIImage {
                 bundle = Bundle.init(path: bundlePath)
             }
         }
-        if bundle != nil {
-            if bundle!.isLoaded == false {
-                bundle?.load()
-            }
-        }
         return UIImage.init(named: name ?? "", in: bundle, compatibleWith: nil)
     }